Start modifications of state handling
Designing the state handling primarily for browser local storage turned
out to be a hindrance when trying to add external storage API support.

The state reducers had side effects that needed to be moved out of the
reducers and into the action creators. Fixing this has exposed flaws
in test design as the tests were reliant on either the side effects or
the synchronous nature of loading the initial state.

At this point, many tests are broken or useless.